% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUSENAS.R
\name{SUSENAS}
\alias{SUSENAS}
\title{Create National Socio-Economic Survey Data Collection Indonesia (SUSENAS)}
\usage{
SUSENAS(data)
}
\arguments{
\item{data}{dataframes}
}
\value{
A dataframes of National Socio-Economic Survey Data Collection Indonesia (SUSENAS)
}
\description{
Create National Socio-Economic Survey Data Collection Indonesia (SUSENAS)
}
